+++ src/main/java/com/harujioh/easing/EaseInQuintEasing.java
@@ -14,7 +14,7 @@ public class EaseInQuintEasing extends BaseEasing {
 	 * @param valueRange
 	 * @param duration
 	 */
-	public EaseInQuintEasing(int startValue, int valueRange, int duration) {
+	public EaseInQuintEasing(double startValue, double valueRange, double duration) {
 		super(startValue, valueRange, duration);
 	}
