+++ src/main/java/com/harujioh/easing/EaseInSineEasing.java
@@ -14,7 +14,7 @@ public class EaseInSineEasing extends BaseEasing {
 	 * @param valueRange
 	 * @param duration
 	 */
-	public EaseInSineEasing(int startValue, int valueRange, int duration) {
+	public EaseInSineEasing(double startValue, double valueRange, double duration) {
 		super(startValue, valueRange, duration);
 	}
